disables support for idempotency key
The cache supporting idempotent requests in the API is not invalidated
by global state changes and thus ends up caching a variety of request
failures that can be fixed by the user.

This results in scenarios where the user believes they have fixed the
conflict but are unable to proceed until the cache has expired.

To re-enable use of the idempotency key, it is recommended to add retry
logic for error responses that have the Idempotency-Replay header set to
true.